Fix `callable_expression` reference (#36)
This PR fixes the `callable_expression` syntax to be able to parse
generic callable expressions such as
`BufferedChannel<*>::registerSelectForSend`. As pat of this PR we
properly fix the issue of ambiguity between comparison expression and
generic calls, e.g. `a<b>(c)` which should be parsed as a call and not
two nested comparison expressions.