Fix parsing `!is` operator (#45)
The `!is` operator can be used in `when` expressions a negative instance
of check:
```
when (s) {
   !is String -> true
}
```

The problem is that here `!` conflicts with the unary `!` operators,
leading to a parse error in cases like the following:
```
when (s) {
   !isString && otherCondition -> true
}
```
Here `!is` should not be parsed as `!is`, rather, it should be parsed as
`! isString`. Unfortunately, there is no straightforward way to handle
this in the tree-sitter grammar, for more info, see
https://github.com/tree-sitter/tree-sitter/issues/4143

The solution here is that `!is` cannot be reported as an operator if
it's immediately followed by an alpha character. We need to handle this
logic in the external scanner.