Newline inside brackets (#48)
This PR simplifies the newline handling inside brackets. In Kotlin,
newline are always allowed inside `()` and `[]`, but not inside `{}`. We
handle the newline explicitly in the grammar rule inside `{}` but for
other cases we simulate a lexer mode: we keep a stack of opening and
closing pairs and interpret the newline character differently based on
the context. This is done by defining an extra symbol `$._newline`,
which is also an external symbol and is handled in the external scanner.

This allows us to parse multi-line annotations like the following:
```
@[
MyQualifier
Target(
    AnnotationTarget.FUNCTION, AnnotationTarget.VALUE_PARAMETER, AnnotationTarget.FIELD,
    AnnotationTarget.PROPERTY
)
Retention(AnnotationRetention.RUNTIME)
]
annotation class MyAnnotation

```