Allow newlines before '(' in call expressions in when conditions (#53)
In Kotlin, newlines are not allowed before `(` in call expressions.
However, there is an exception in when conditions: newlines are allowed
there! This is not specified in the Kotlin specifications but the Kotlin
compiler behaves like this.

To fix this, we have to copy the expression nonterminal in
`when_condition` and change the `call_expression` node with another one
that allows newlines. Also, changes should be made in the lexer to allow
this.